+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Version History
 
-## 0.5.0 (unreleased)
+## 0.6.0 (unreleased)
+
+## 0.5.0
 
 **Breaking changes**